search: don't send repo count if count==0
With this change we omit `RepositoriesCount` from the final progress
event if the count equals 0. The client doesn't display the repository
count if we don't send it, which makes more sense than displaying `0
repositories`.

This also buys us some time while we work on a good solution of how to
efficiently return the number of repos in the global scope of the user.

Without this check it is possible that we report `500+ matches in 0.1s
from 0 repositories` if a global search was too fast.