API docs: LSIF: adjust definition ranges / fix incorrect references
Some usage examples / references were incorrectly reporting other
random symbols:

<img width="929" alt="image" src="https://user-images.githubusercontent.com/3173176/127557175-ec2b88a2-da79-4f8c-bc7c-a5ccdafce597.png">

After investigating, the `DocumentationDefinition` request definitely produces the right
location for the definition - exactly the same location that we send in a regular `graphql?References`
request - which introduced the question "why are our references wrong, then?"

The only difference is that in the definition resolution code in `DocumentationReferences`,
I did not previously `adjustLocations`. That seems obviously wrong, and a likely cause - so
here I fix that.

I suspect it'll fix the issue, but I don't have confirmation yet.

Fixes #23376

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>