metrics: Adjust src_access_granted_private_repo
Instead of counting the number of repos when listing, we just need to
increase the count by 1. The intention of this metric for now is to get
a feel for how often we'd need to log access granted events and we only
have one event per access, not one event for each repo.