search: Fix duplicate React key error in search reference
I must have forgotten to update the condition in one of the refactors.
Currently this creates a lot of `Warning: Encountered two children with the same key, 0.`
errors because every example that is like `{test}` will contain an empty
token that starts at the same position as the placeholder token.