add Metadata method to RepoStore
This adds a method to fetch repo metadata for search results. It only
returns the subset of fields needed for decorating search results.

This adds a field JoinGitserverRepos to the ReposListOptions struct,
which forces a join on the gitserver_repos table so fields from that
table are select-able. This allows us to do a single query for all repo
metadata rather than two separate queries.