handle datbase I/O timeout
With this change we return a timeout-alert instead of an plain error in
case of db I/O timeout.

On main we reliably get an I/O Error from the db for extremely short
timeouts, like `router timeout:1ms index:no`. Currently we don't handle
the error and simply surface it to the user.