Tweak the page title for code intelligence settings.
Previously, the page title was "404 not found" while the code
intelligence settings UI was loading in the browser. The title would
then get updated to the correct title after the SPA finished loading.
This commit adds a server-side route handler so that the title is "Code
intelligence" instead of "404 not found" while the page is loading.

We also remove the "Precise" prefix from the title in code intel
settings pages.  The "Precise" prefix is technically accurate, but it's
inconsistent with the "Code Intelligence" button that users click to view
those settings. "Precise code intelligence" is also so long that it's
usually truncated by the browser so that you can't see the
config/uploads/auto-index suffix in the subpages.