handle database I/O time out (#23478)
With this change we return a time out alert instead of a plain error in
case of a db i/o time out.

Locally, we reliably get an I/O Error from the db for extremely short
timeouts, like `router timeout:1ms index:no`. Currently we don't
handle the error and simply surface it to the user. 

This is unlikely to occur for a normal user, but the integration tests
are flaky because of that.