search: OR public and private queries
The current implementation breaks sorting if the user has access to
private repositories. Since the queries for public and private repos run
independently, sorting applies to the individual streams which can lead
to private repos appearing on top of public repos.

With this change we `OR` the two queries. This preserves the benefit of
not having to resolve large lists of public repositories before sending
back results.

We have to monitor a potential impact on performance though. Zoekt will
translate the query to a much larger matchtree which could potentially
impact performance.

I will monitor the sentinel queries and make an assessment.