Don't expire batch specs if batch spec execution points to them
This fixes #23502 but not in the way I originally thought I would fix
it.

My first idea was to make the `DELETE` cascade and also delete the
`batch_spec_execution` when the `batch_spec` is deleted.

But then I was unsure whether that is what we want. Since we'll have a
list view of executions (which we don't have for batch specs) I'm not
sure whether we want to delete those.

So I'm deferring on that and instead changed the code to not delete
batch specs if they're attached to a batch spec execution.

I think that's fine for now and then we can think about how we want to
clean up batch spec executions.

I also changed `CreateBatchSpecExecution` to set `batch_spec_id` (if
set) because that cost me 30min of debugging my tests.

And in order to get this working in the tests I had to make the foreign
key constraint deferrable. That's what the migration is for.