httpcli: Expand usage across the codebase
This commit accomplishes three things:

1. Use the httpcli package for internal HTTP clients.
2. Use the httpcli for some remaining external use cases that weren't using it.
3. With retries in place, remove the `WaitForX` anti-pattern.

Why is `WaitForX` an anti-pattern? Because in production availability
isn't constant, and we should instead structure the code to retry when
needed, which this commit achieves.

I noticed that when repo-updater and gitservers were deployed at the
same time, repo-updater would go into a crash loop since it couldn't
wait for all gitservers to be up.