add more repo metadata (#23493)
* make getEventRepoMetadata an independent function

It only uses the database from the receiver, so just pass that in
explicitly.

* move metadata function to a new file

Metadata population feels slightly unrelated to the rest of the file, so
split it off.

* add Metadata method to RepoStore

This adds a method to fetch repo metadata for search results. It only
returns the subset of fields needed for decorating search results.

This adds a field JoinGitserverRepos to the ReposListOptions struct,
which forces a join on the gitserver_repos table so fields from that
table are select-able. This allows us to do a single query for all repo
metadata rather than two separate queries.

* add repoLastFetched to search events

This adds the last fetched time to all search result events.

It also slightly rearranges the construction of events as well to avoid
pre-declaring a bunch of variables.

* add private to repo event

* add fields to typescript types

* review feedback