internal/database: Ensure only user added private repos are shown
The authzQuery had an OR condition on the query to check if the repo
is owned by the authenticated user. This commit changes it to an AND
condition to ensure that if a private repo was not added by the
currently authenticated user, they don't see it in their search
results.