Make calls to sentry.Init explicit, make conf.Init func visible
Alright, how did we get here?

We found error messages in `executor` (deployed outside of a Sourcegraph
cluster) that said it can't connect to
`http://sourcegraph-frontend-internal` to fetch the connection.

Why would `executor` even try to fetch the connection?

Thanks to `./dev/depgraph` we found out that `executor` uses `trace` and
that imports `sentry` and that imports `conf` and `conf` has an `init()`
function that starts a goroutine that watches for configuration changes
with `conf.Watch`.

So, to :fingerquotes: improve :fingerquotes: on that we did a couple of things here:

1. Pull out the `sentry.Init()` calls from `trace.Init()` so they're
   visible.
2. Change `conf.init` into capital-case `Init()` and call that
   everywhere we suspect a `httpcli.ExternalClient` is used (because
   that's what the `Init` function updates)
3. Do not call any of that in `executor`

At this point I want to say "I don't know good from evil anymore", but I
do know that `conf` is evil, so I don't know, man