API docs: fix needless rerendering of page
This was incredibly difficult for me to track down, probably 9h of my time overall.
See #21200 for details.

We'll need to apply this same fix to other pages in this `routes.tsx` file, because
they are 100% affected by this. Also, this isn't a complete fix - there are other
issues (like not using `React.memo` on this page) that mean typing in the search bar
is still incredibly sluggish.

Helps #21200

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>