share a max payload size between client and server
Previously, if the writer wrote an event that was greater than the max
buffer size of the client, it would fail to parse and stop processing
further events. This can happen when we match a very large line, or have
many matches on a single line. Instead, if the writer shares the same
max event size, it can just ignore files matches that are too large, log
a message, and continue sending results.