repo-updater: Run repo perms sync if a repo is added or modified
In this commit we make an API change on the repo syncer so that it can
trigger a repo permissions sync without needing access to the
permissions syncing client. `repos.Syncer.PermsSyncer` interface only
defines the `ScheduleRepos` method because the syncer does not need
access to the `ScheduleUsers` method from the `PermsSyncer` yet.

Also, when a new repo is created or an existing one is modified we
trigger a repo permissions sync if the final state of the repo at the
end of the create / update operation indicates that it is private.