internal/repos: Refactor schedule repo perms sync decision
In this commit we make the decision to schedule a permissions sync at
the time of repo creation / update. This is necessary, because the
`stored` repo object gets updated with the new state upon a successful DB
write and leaves us with no information on the previous state of the
repo prior to the DB write.

Additionally, using a single flag simplifies the logic to trigger repo
perms sync for both the cases where a new private repo is added and
where an existing public repo is made private.