search: bring back repo count for fast global queries (#23498)
This brings back the repo count for those queries that are faster than repo
resolution.

See https://github.com/sourcegraph/sourcegraph/pull/23366 for more background.

The idea is to use repository resolution as is, but unlike now we let it finish
even if we hit a limit early. To make this work, we introduce a wrapper for
contexts that let's us communicate the reason for a cancellation. Repo
resolution in turn ignores cancellation if the context was cancelled because we
hit a limit.