API docs: sidebar: clicking "Constants", "Variables", etc. behave as buttons
Previously, the dropdown button and `Constants` text in `> Constants` were two
seperate functions: one expands the area, the other jumps to the "Constants"
section of the page. This is a bit confusing, though, because jumping to the
section of the page _also_ means it gets automatically expanded. This makes
some users think clicking the text makes the area expand, which is not true.

So, make clicking on "Constants", "Variables", etc. do what the user expects:
expand/collapse.

The only caveat to this is that _some expandable sections_, like `type Foo struct`
with methods below it, will continue to act as having two seperate buttons.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>