repo-updater: Run repo perms sync if a repo is added or modified (#23721)
In this commit we make an API change in the repo syncer so that it can
trigger a repo permissions sync without needing access to the
permissions syncing client. The `repos.Syncer.PermsSyncer` interface
defines just the `ScheduleRepos` method because the syncer does not
need access to the `ScheduleUsers` method from the `PermsSyncer` yet.

Here, we make the decision to schedule a permissions sync at
the time of repo creation / update. This is necessary, because the
`stored` repo object gets updated with the new state upon a successful DB
write and leaves us with no information on the previous state of the
repo prior to the DB write.

Additionally, using a single flag simplifies the logic to trigger repo
perms sync for both the cases where a new private repo is added and
where an existing public repo is made private.

Finally, we also add adequate integration tests for `syncer.SyncRepo`
and `syncer.SyncExternalService` which happen to be the two call sites
that invoke `syncer.sync` - the method ultimately responsible for
conditionally scheduleing a repo permissions sync request.

Co-authored-by: Ryan Slade <ryanslade@gmail.com>
Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>