authz/perforce: support wildcard permissions (#23755)
Prefix queries were implemented with PostgreSQL’s `LIKE`, so we can leverage the `%` syntax currently used to support prefixes to mimic `…` wildcards along a path. To support `*` wildcards, which only match a single directory, we switch the prefix queries to use PostgreSQL's `SIMILAR TO`, which allows us to leverage some simple regex, namely `[^/]+`. Together, we get directory-level support for wildcard paths in Perforce permissions tables.

To reflect the expanded functionality of `XXXPrefixes` and friends, prefixes are now referred to as `XXXContains` in relevant code. The concept of `Contains` being a prefix is now in the Perforce permissions syncer instead of being in database.