aggregate line matches in searcher (#23918)
Previously, searcher was creating a new LineMatch for every fragment
matched. For long lines with many matched fragments, this meant
duplicating the content of that line many times in both allocations and
when serializing. This was causing significant memory issues in searcher
now that we don't have strict limits on the number of matches per file.