all: instrument version in trace and prometheus (#24127)
We embed a version identifier in our binaries. This adds it to both
tracing and metrics. For tracing we add the tracer tag
"service.version". For metrics we add the gauge "src_version" which has
a label version. For metrics this can then be joined against[1].

This is useful for observing the impact of a version change on
metrics. Additionally when receiving debug information from customers
this helps us understand exactly which versions are being run.

[1]: https://www.robustperception.io/exposing-the-software-version-to-prometheus