internal/extsvc: Use stricter check for equality in CodeHostOf
In this commit we extract the codehost URL from the given repo name
and apply an exact case-insensitive match instead of only checking for
the prefix. The previous approach does not handle corner cases like:
github.com.example.com/foo/bar and falsely identifies it as a repo
belonging to the codehost with URL github.com.