nix: a shell.nix for running go test (#24230)
This is a first step towards support nix in our development
environment. It only covers go test and yarn. I am still a nix noob as
well, so could be much better.

Different to our usual environment is that it runs postgres under
~/.sourcegraph with a unix socket. No need to manage service. Must
remember to run pg_ctl shutdown if you want to stop.