remove ineffective theme-class wrappers for text emphasis
The intent of this is to "Make sure text color utilities are theme-aware". However, `$theme-colors` only contains a color named `merged` (purple, for showing merged PRs) whose value is a SCSS constant, not a CSS variable, which means it can't change between the light and dark themes anyway. Therefore, there is no need to make this theme-aware.

If the intent was to make all text color utilities (eg `.text-secondary`, `.text-success`, etc.) theme-aware, then this approach would also not work without significant additional complexity because Bootstrap expects the colors to be SCSS constants so that it can do CSS color math on them to make border/hover/etc. variants from the base color for buttons and other components.