internal: Add logging to observe rate limits of external services
Recently we've noticed that we see frequent "internal rate limit
exceeded" errors on dotcom. Whereas, based on the investigation of the
current site-config and the external service config of the cloud
default GitHub external service, the rate limit should be
infiinite (math.MaxFloat64) as we don't set the `requestsPerHour`
config at all. These logs will help us to get an insight into the
current state of the rate limiter for these external services.