database: parameterise indexable repos min stars
This commit introduces an environment variable
`SRC_INDEXABLE_REPOS_MIN_STARS` that can be used to change the minimum
number of stars neeeded for a public repo to be indexed by Zoekt on
sourcegraph.com.

Additionally, we change the threshold from 13 to 8, in order to grow the
index to ~1.5M repos.