repo-updater: lower burst for rate limited test
A bug was fixed in the rate limit library. That fix caused the "not
enough quota available" test to start failing. I believe the issue is
the bug fix made the rate limit correctly respect bursts for WaitN, so
we actually had enough tokens due to bursting. This commit lowers the
burst so we don't have enough tokens after 1s.

The fix in the rate limit library is here:
https://github.com/golang/time/commit/1f47c861a9ac5a6e7645609f91b895398ff31642

The parent of that commit has the tests passing without this change.