remove needless ~ from CSS/SASS @imports
Previously, Webpack's sass-loader (incorrectly) required a `~` before a CSS/SASS `@import` for the path to be resolved (i.e., to refer to a path in `node_modules`). This was incorrect because SASS's specification says that *all* non-relative `@import` paths should be resolved.

Now Webpack's sass-loader no longer needs `~`, and in fact it recommends against using it (see https://webpack.js.org/loaders/sass-loader/#resolving-import-at-rules). This removal has no effect on the bundle.