remove import cycle and needless exports around eventLogger
Moved 2 functions (`pageViewQueryParameters and `handleQueryEvents`) from `analyticsUtils.tsx` to `eventLogger.ts` because they were only used in the latter.

This also removes an import cycle that messed up esbuild. (Perhaps esbuild's behavior is buggy here, but either way, this is a good change because it removes needless export API surface.)