httpcli: only retry DNS lookup failures 3 times
We already had this check, but it expected the error type to be
"*url.Error". In practice the error I see is a "*net.OpError" which
wraps the core "*net.DNSError" in the case of DNS lookup failures. From
following the go stdlib code "no such host" errors always originate via
a "*net.DNSError". So instead we do error unwrapping to check against
that.

In practice without the smaller retries we get elevated search response
times if any zoekt replica is unavailable.