httpcli: log to span if retry policy called
During a recent site incident we had some spans take a very long
time. The root cause turned out to be the automatic retries we do. This
adds visibility into our traces when this occurs.

Additionally we add a prometheus counter which is incremented every time
the retry policy decides to retry.