httpcli: only log to trace if retrying
I didn't realise, but the retry policy function is always called even
after a successful request. To avoid trace log spam, I've updated the
tracing to only occur if we decide to retry or if we have done more than
one attempt. Additionally I've updated the event name to be clearer.