support fetching a subset of extension manifest fields for faster loads
This cuts the `graphql?Extensions` response size from ~800kb to ~5kb (compressed), or ~1.3MB to ~86kb (uncompressed). Also, it eliminates the need to JSONC-parse the result (which is slower than parsing JSON), so there's also a significant decrease in CPU needed.

Previously, the `graphql?Extensions` result was very large and slow to load/parse (~800kb compressed, ~1.3MB uncompressed, for the default set of extensions) because it included the full JSON `package.json` manifest for each extension. This file contains base64-encoded images and the full Markdown README contents, neither of which are necessary to execute extensions.

Now, only a subset of the manifest's fields are fetched (`url`, `contributes`, `activationEvents`). TypeScript types are used to enforce that no callers are using fields that aren't fetched.