Don't infer lsif-java auto-index config for non-indexable repos
Previously, we inferred lsif-java indexing configuration for repos that
didn't have any Java or Scala source files. This caused indexing jobs to
fail when we shouldn't have indexed the repo in the first place. This
commit fixes that issue by requiring the repo to contain at least a
single Java or Scala file in order to infer the lsif-java indexing
configuration.

The refactoring for this commit also made it easy to fix another bug
where lsif-java would fail to index repos that contained both a toplevel
`pom.xml` file and `lsif-java.json` file (because it ambiguous which
    build tool to use). Now, we use the `lsif-java index --build-tool=NAME` to
explicitly set the build tool to LSIF when there is an `lsif-java.json`
file.