ask Google, etc. not to index useless pages
Introduces per-route tagging of either `index` or `noIndex`. When `noIndex`
is specified, we respond with a `X-Robots-Tag: noindex` header instructing
Google etc. not to index the page.

I have manually vetted / tagged each route with what I think is most appropriate,
leaning on the side of _not indexing_ in general (because indexing bad content
is _really_ bad for us, much worse than erring on the side of indexing.)

We can continue to revise individual routes, etc. as needed.

Fixes #24696

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>